Started FillInTiles function
Finished most of the FillInTiles function. The idea behind this function is that a tile should be filled in with a number in one of two cases:
1: the only item in "possibleNotes" for that tile is the number.
2: if a number appears only once in the possibleNotes (with respect to its row, column or 3x3 square), the tile that holds that number in its possibleNotes must be that number.

1 is implemented quickly (it was very easy to do). 2 is almost done, so far it only checks with respect to the row and column that the tile is in (I still need to check by 3x3 square).

Currently, despite only one analysis algorithm being completed (which I call "general analysis") and not completing the FillInTiles function, the example board in "input.txt" can be completed in 4 loops. (Hence why the entire program is in a for loop that runs 4 times) Harder Sudoku puzzles likely cannot be solved with the minimal work I have put in so far. (But there is more to come!)