Finished FillInTiles function
FillInTiles was a massive pain, but v1 is finally done! FillInTiles can be massively optimized, and I plan to do that in the future. The current program will run through as many times as it takes to solve the puzzle. Therefore, if the puzzle is too difficult for this program, it will just continually run. I am using the android app called: "Sudoku - Classic Sudoku Puzzle" by Guru Puzzle Game to find inputs and test the program. So far, this can solve easy, medium, and hard puzzles from the app, but started to struggle when I tried expert puzzles.